{"schema":"decodex/commit/1","summary":"Extract Swift scroll capture bridge owner","authority":"manual"}
