+++ docs/reference/workspace-layout.md
@@ -309,8 +309,11 @@ The main host-kit files are split by responsibility:
 - `NativeHostImageBridge.swift`: RGBA/CoreGraphics image conversion used by the FFI bridge
 - `Sources/RsnapHostBridge/HostSessionFFI.swift`: Swift bridge session protocol models and
   session-handle adaptation used by the native host
-- `Sources/RsnapHostBridge/HostFFI.swift`: Swift bridge scroll and remaining non-frozen FFI
-  surfaces used by the native host
+- `Sources/RsnapHostBridge/HostFFI.swift`: shared RGB/RGBA models plus selection-transform,
+  auto-center, and BGRA frame-sampling bridge surfaces used by the native host
+- `Sources/RsnapHostBridge/ScrollCaptureFFI.swift`: Swift bridge scroll minimap planning,
+  scroll-capture observation models, session-handle lifecycle, and stitched preview/export image
+  adaptation
 - `Sources/RsnapHostBridge/ExportEncoderFFI.swift`: Swift bridge PNG encoding, frozen display
   crop, mosaic privacy patch, and frozen overlay export-image adaptation
 - `Sources/RsnapHostBridge/LiveSamplerFFI.swift`: Swift bridge live sampler models,
